feat(stdlib): Add `Path.updateExtension`
This adds `Path.updateExtension` so users can set the current extension.

Edge Cases:
* Regular, we just chop the extension and then append `.<ext>` to the end
* Directory, we don't set the extension
* Empty extension sets something like `file.ext` to `file.`, this seems weird but I think it's the best option if a user actually wants to clear the extension they should use `removeExtension`.

Note: One thing I noticed is our examples for `Path` omit the `Path.` which means unless you do` use Path.*` (which I don't think we encourage) the examples are not runnable off the bat, I think we should fix this in a seperate pr for consistency across the stdlib.